[mesut] updated with respect to @jmoody comments
