--- .ruby-version
@@ -1,1 +1,0 @@
-2.3.0
