+++ Gemfile
@@ -1,5 +1,5 @@
-# A sample Gemfile
 source "https://rubygems.org"
 
-gem 'calabash-cucumber', '0.20.5'
-gem 'calabash-android', '0.9.0'
+
+gem 'calabash-cucumber', '>= 0.21', '< 1.0'
+gem 'calabash-android', '>= 0.9', '< 1.0'
