+++ features/002_register.feature
@@ -7,6 +7,7 @@ Feature: Register
     Given I prepare the state of the app as logout user
     Given I am in the register page
 
+    # Scenario outline is not implemented in App Center!
     # Scenario Outline: Obtaining more information
     #   When I enter <First Name> <Last Name> <Email> and <Password> for registration
     #   Then I should see the <Result>
