+++ features/step_definitions/login.rb
@@ -14,13 +14,13 @@ Given(/^I prepare the state of the app as logout user$/) do
 end
 
 When(/^I go to my profile$/) do
-	@current_page = page(ProductsPage).await(timeout: 30)
-	@current_page.go_to_profile
+  @current_page = page(ProductsPage).await(timeout: 30)
+  @current_page.go_to_profile
 end
 
 When(/^I log out$/) do
-	@current_page = page(ProfilePage).await(timeout: 30)
-	@current_page.logout
+  @current_page = page(ProfilePage).await(timeout: 30)
+  @current_page.logout
 end
 
 When(/^I log in$/) do
@@ -28,22 +28,22 @@ When(/^I log in$/) do
 end
 
 When(/^I touch the log-in button$/) do
-	@current_page.touch_login_button
+  @current_page.touch_login_button
 end
 
 Given(/^I am in the login page$/) do
-	@current_page = page(LoginPage).await(timeout: 30)
+  @current_page = page(LoginPage).await(timeout: 30)
 end
 
 When(/^I fill in "(.*?)" as "(.*?)"$/) do |input, text|
-	@current_page.fill_in_input(input, text)
+  @current_page.fill_in_input(input, text)
 end
 
 When(/^I touch the {string} button$/) do |text|
-	touch_button_text text
+  touch_button_text text
 end
 
 Then(/^I should see "(.*?)" text$/) do |text|
-	wait_until_text_visible text
-	check_element_exists "* text:'#{text}'"
+  wait_until_text_visible text
+  check_element_exists "* text:'#{text}'"
 end
